meeting_server/requirements.txt
Pillow
Django
drf-writable-nested
django-phonenumbers-field
django-phonenumbers
djangorestframework